auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
redis==5.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
email-validator==2.2.0
